[feat] separate sockets out of the main server file
